+++ src/app.controller.ts
@@ -1,12 +1,10 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
-  constructor(private readonly appService: AppService) {}
 
   @Get()
-  getHello222(): string {
-    return this.appService.getHello();
+  getHello(): string {
+    return 'Hello World';
   }
 }
