+++ src/resources/boards/dto/create-board.dto.ts
@@ -1,4 +1,4 @@
-import { ColumnEntity } from '../../entities/column.entity';
+import { ColumnEntity } from '../../../entities/column.entity';
 
 export class CreateBoardDto {
   id?: string;
