+++ src/resources/boards/entities/board.entity.ts
