+++ src/resources/users/entities/user.entity.ts
