+++ src/resources/users/users.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from '../entities/user.entity';
-import { TaskEntity } from '../entities/task.entity';
+import { UserEntity } from '../../entities/user.entity';
+import { TaskEntity } from '../../entities/task.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity, TaskEntity])],
